Fix bug that select response is discarded
Added setup_callback() before sending select request.
