+++ sdk/modules/lte/altcom/api/socket/altcom_select.c
@@ -43,6 +43,7 @@
 #include "altcom_sock.h"
 #include "altcom_socket.h"
 #include "altcom_select.h"
+#include "altcom_select_ext.h"
 #include "altcom_seterrno.h"
 #include "apicmd_select.h"
 #include "buffpoolwrapper.h"
@@ -57,7 +58,6 @@
 #define SELECT_RES_DATALEN (sizeof(struct apicmd_selectres_s))
 
 #define SELECT_REQ_FAILURE -1
-#define SELECT_ID_GEN_AUTO -1
 
 /****************************************************************************
  * Private Types
@@ -74,11 +74,20 @@ struct select_req_s
   int32_t           timeout;
 };
 
+struct select_asynccb_s
+{
+  int32_t                     select_id;
+  altcom_select_async_cb_t    callback;
+  FAR void                    *priv;
+  FAR struct select_asynccb_s *next;
+};
+
 /****************************************************************************
  * Private Data
  ****************************************************************************/
 
 static int32_t g_select_id = 0;
+static FAR struct select_asynccb_s *g_callbacklist_head = NULL;
 
 /****************************************************************************
  * Private Functions
@@ -93,6 +102,182 @@ static int32_t generate_selectid(void)
   g_select_id++;
 
   return (g_select_id & 0x7fffffff);
+}
+
+
+/****************************************************************************
+ * Name: allocate_callbacklist
+ ****************************************************************************/
+
+static FAR struct select_asynccb_s *allocate_callbacklist(int32_t select_id)
+{
+  FAR struct select_asynccb_s *list;
+
+  /* Allocate callback list */
+
+  list = (FAR struct select_asynccb_s*)BUFFPOOL_ALLOC
+           (sizeof(struct select_asynccb_s));
+  if (!list)
+    {
+      DBGIF_LOG_ERROR("Failed to allocate memory.\n");
+      return NULL;
+    }
+
+  memset(list, 0, sizeof(struct select_asynccb_s));
+
+  /* Initialize cllback list */
+
+  list->select_id = select_id;
+  list->next      = NULL;
+
+  return list;
+}
+
+/****************************************************************************
+ * Name: free_callbacklist
+ ****************************************************************************/
+
+static void free_callbacklist(FAR struct select_asynccb_s *list)
+{
+  (void)BUFFPOOL_FREE((void*)list);
+}
+
+
+/****************************************************************************
+ * Name: add_callbacklist
+ ****************************************************************************/
+
+static void add_callbacklist(FAR struct select_asynccb_s *list)
+{
+  FAR struct select_asynccb_s *list_ptr = g_callbacklist_head;
+
+  if (g_callbacklist_head)
+    {
+      /* Search end of list */
+
+      while(list_ptr->next)
+        {
+          list_ptr = list_ptr->next;
+        }
+      list_ptr->next = list;
+    }
+  else
+    {
+      g_callbacklist_head = list;
+    }
+
+}
+
+/****************************************************************************
+ * Name: delete_callbacklist
+ ****************************************************************************/
+
+static int32_t delete_callbacklist(FAR struct select_asynccb_s *list)
+{
+  FAR struct select_asynccb_s *list_ptr = g_callbacklist_head;
+  FAR struct select_asynccb_s *prev_ptr = NULL;
+
+  if (g_callbacklist_head)
+    {
+      while (list_ptr)
+        {
+          if (list_ptr == list)
+            {
+              /* Check begining of the list */
+
+              if (prev_ptr)
+                {
+                  prev_ptr->next = list_ptr->next;
+                }
+              else
+                {
+                  g_callbacklist_head = list_ptr->next;
+                }
+
+              return 0;
+            }
+          prev_ptr = list_ptr;
+          list_ptr = list_ptr->next;
+        }
+    }
+
+  return -1;
+}
+
+/****************************************************************************
+ * Name: search_callbacklist
+ ****************************************************************************/
+
+static FAR struct select_asynccb_s *search_callbacklist(int32_t select_id)
+{
+  FAR struct select_asynccb_s *list_ptr = g_callbacklist_head;
+
+  while (list_ptr)
+    {
+      if (list_ptr->select_id == select_id)
+        {
+          return list_ptr;
+        }
+      list_ptr = list_ptr->next;
+    }
+
+  return NULL;
+}
+
+/****************************************************************************
+ * Name: setup_callback
+ ****************************************************************************/
+
+static void setup_callback(int32_t select_id, altcom_select_async_cb_t cb,
+                           FAR void *priv)
+{
+  FAR struct select_asynccb_s* list;
+
+  /* Check if select ID is already in use */
+
+  if (!search_callbacklist(select_id))
+    {
+      list = allocate_callbacklist(select_id);
+      if (list)
+        {
+          list->callback = cb;
+          list->priv     = priv;
+
+          /* Add list to the end of list */
+
+          add_callbacklist(list);
+        }
+    }
+  else
+    {
+      DBGIF_LOG1_WARNING("This select ID[%d] already in use.\n", select_id);
+    }
+}
+
+/****************************************************************************
+ * Name: teardown_callback
+ ****************************************************************************/
+
+static void teardown_callback(int32_t select_id)
+{
+  FAR struct select_asynccb_s *list;
+  int32_t                     ret;
+
+  list = search_callbacklist(select_id);
+  if (list)
+    {
+      /* Delte list from callback list */
+
+      ret = delete_callbacklist(list);
+      if (ret < 0)
+        {
+          DBGIF_LOG_ERROR("Failed to delete callbck list.\n");
+        }
+      else
+        {
+          free_callbacklist(list);
+        }
+    }
 }
 
 /****************************************************************************
@@ -165,11 +350,6 @@ static int32_t select_request(FAR struct select_req_s *req)
     {
       err = ALTCOM_ENOMEM;
       goto errout;
-    }
-
-  if (req->select_id == SELECT_ID_GEN_AUTO)
-    {
-      req->select_id = generate_selectid();
     }
 
   /* Fill the data */
@@ -270,11 +450,6 @@ static int32_t select_request_async(FAR struct select_req_s *req)
       goto errout;
     }
 
-  if (req->select_id == SELECT_ID_GEN_AUTO)
-    {
-      req->select_id = generate_selectid();
-    }
-
   /* Fill the data */
 
   fill_select_request_command(cmd, req);
@@ -310,82 +485,6 @@ errout:
  ****************************************************************************/
 
 /****************************************************************************
- * Name: altcom_select_request_asyncsend
- ****************************************************************************/
-
-int altcom_select_request_asyncsend(int maxfdp1, altcom_fd_set *readset,
-                                    altcom_fd_set *writeset,
-                                    altcom_fd_set *exceptset)
-{
-  int32_t             ret;
-  int32_t             result;
-  struct select_req_s req;
-
-  /* Check Lte library status */
-
-  ret = altcombs_check_poweron_status();
-  if (0 > ret)
-    {
-      altcom_seterrno(-ret);
-      return -1;
-    }
-
-  req.select_id = SELECT_ID_GEN_AUTO;
-  req.maxfdp1   = maxfdp1;
-  req.request   = APICMD_SELECT_REQUEST_BLOCK;
-  req.readset   = readset;
-  req.writeset  = writeset;
-  req.exceptset = exceptset;
-  req.timeout   = SYS_TIMEO_FEVR; /* ignore when do select_request_async */
-
-  result = select_request_async(&req);
-
-  if (result == SELECT_REQ_FAILURE)
-    {
-      return -1;
-    }
-
-  return result;
-}
-
-/****************************************************************************
- * Name: altcom_select_cancel_request_send
- ****************************************************************************/
-
-int altcom_select_cancel_request_send(int id)
-{
-  int32_t             ret;
-  int32_t             result;
-  struct select_req_s req;
-
-  /* Check Lte library status */
-
-  ret = altcombs_check_poweron_status();
-  if (0 > ret)
-    {
-      altcom_seterrno(-ret);
-      return -1;
-    }
-
-  req.select_id = id;
-  req.maxfdp1   = 0;
-  req.request   = APICMD_SELECT_REQUEST_BLOCKCANCEL;
-  req.readset   = NULL;
-  req.writeset  = NULL;
-  req.exceptset = NULL;
-  req.timeout   = SYS_TIMEO_FEVR; /* ignore when do select_request_async */
-
-  result = select_request_async(&req);
-
-  if (result == SELECT_REQ_FAILURE)
-    {
-      return -1;
-    }
-
-  return result;
-}
-
-/****************************************************************************
  * Name: altcom_select_nonblock
  ****************************************************************************/
 
@@ -407,7 +506,7 @@ int altcom_select_nonblock(int maxfdp1, altcom_fd_set *readset,
       return -1;
     }
 
-  req.select_id = SELECT_ID_GEN_AUTO;
+  req.select_id = generate_selectid();
   req.maxfdp1   = maxfdp1;
   req.request   = APICMD_SELECT_REQUEST_NONBLOCK;
   req.readset   = readset;
@@ -446,7 +545,7 @@ int altcom_select_block(int maxfdp1, altcom_fd_set *readset,
       return -1;
     }
 
-  req.select_id = SELECT_ID_GEN_AUTO;
+  req.select_id = generate_selectid();
   req.maxfdp1   = maxfdp1;
   req.request   = APICMD_SELECT_REQUEST_BLOCK;
   req.readset   = readset;
@@ -493,6 +592,123 @@ int altcom_select(int maxfdp1, altcom_fd_set *readset,
   if ((ret == SELECT_REQ_FAILURE) && (altcom_errno() == ALTCOM_ETIMEDOUT))
     {
       ret = 0;
+    }
+
+  return ret;
+}
+
+/****************************************************************************
+ * Name: altcom_select_async
+ ****************************************************************************/
+
+int altcom_select_async(int maxfdp1, altcom_fd_set *readset,
+                        altcom_fd_set *writeset, altcom_fd_set *exceptset,
+                        altcom_select_async_cb_t callback, void* priv)
+{
+  int ret;
+  struct select_req_s req;
+
+  if (!callback)
+    {
+      altcom_seterrno(ALTCOM_EINVAL);
+      return -1;
+    }
+
+  /* Check LTE library status */
+
+  ret = altcombs_check_poweron_status();
+  if (0 > ret)
+    {
+      altcom_seterrno(-ret);
+      return -1;
+    }
+
+  req.select_id = generate_selectid();
+  req.maxfdp1   = maxfdp1;
+  req.request   = APICMD_SELECT_REQUEST_BLOCK;
+  req.readset   = readset;
+  req.writeset  = writeset;
+  req.exceptset = exceptset;
+  req.timeout   = SYS_TIMEO_FEVR; /* ignore when do select_request_async */
+
+  setup_callback(req.select_id, callback, priv);
+
+  ret = select_request_async(&req);
+
+  if (ret == SELECT_REQ_FAILURE)
+    {
+      teardown_callback(req.select_id);
+      return -1;
+    }
+
+  return ret;
+}
+
+
+/****************************************************************************
+ * Name: altcom_select_async_exec_callback
+ ****************************************************************************/
+
+int altcom_select_async_exec_callback(int32_t id, int32_t ret_code,
+                                      int32_t err_code,
+                                      altcom_fd_set *readset,
+                                      altcom_fd_set *writeset,
+                                      altcom_fd_set *exceptset)
+{
+  int ret = -1;
+  FAR struct select_asynccb_s *list;
+
+  list = search_callbacklist(id);
+  if (list)
+    {
+      /* execute callback */
+
+      list->callback(ret_code, err_code, id, readset, writeset, exceptset,
+                     list->priv);
+
+      teardown_callback(id);
+      ret = 0;
+    }
+
+  return ret;
+}
+
+/****************************************************************************
+ * Name: altcom_select_async_cancel
+ ****************************************************************************/
+
+int altcom_select_async_cancel(int id, bool is_send)
+{
+  int ret = 0;
+  struct select_req_s req;
+
+  teardown_callback(id);
+
+  if (is_send)
+    {
+      /* Check LTE library status */
+
+      ret = altcombs_check_poweron_status();
+      if (0 > ret)
+        {
+          altcom_seterrno(-ret);
+          return -1;
+        }
+
+      req.select_id = id;
+      req.maxfdp1   = 0;
+      req.request   = APICMD_SELECT_REQUEST_BLOCKCANCEL;
+      req.readset   = NULL;
+      req.writeset  = NULL;
+      req.exceptset = NULL;
+      req.timeout   = SYS_TIMEO_FEVR; /* ignore when do select_request_async */
+
+      ret = select_request_async(&req);
+
+      if (ret == SELECT_REQ_FAILURE)
+        {
+          return -1;
+        }
     }
 
   return ret;
